backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
yfinance>=0.2.36
polars>=0.20.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dateutil>=2.8.2
websockets>=12.0
